#1 Add auto text resize for text fields NPC sheet
The font size automatically resizes when more text is added to the line
than it is supposed to be in the default font size.

This has its limits when even Acrobat thinks the font is too small. Then
still a plus sign may be shown in order to visualize that there is too
much text in the form box.